cancerscreening: Streamline Access to Cancer Screening Data

cancerscreening provide a easy way to download cancer screening data from the Kenya Health Information System (KHIS) using R.

Most function begin with the prefix get_ followed by the screening area cervical or breast. The Goal is to allow the download of data associated with the data of interest, e.g. get_cervical_screened, get_cervical_positive, or get_cervical_treated.

cancerscreening is "pipe-friendly" and, in fact re-exports %>% but does not require its use.

Please see the cancerscreening website for full documentation:

https://cancerscreening.damurka.com/index.html

In addition to function-specific help, there are several articles which are indexed here:

Article index

See also

Author

Maintainer: David Kariuki hello@damurka.com (ORCID) [copyright holder]