Retrieve and Format Breast Cancer Screening Data

.get_breast_data() retrieves breast cancer screening data for a specified period from the KHIS API server using get_analytics().

Usage

.get_breast_data(
  element_ids,
  start_date,
  end_date = NULL,
  level = c("kenya", "county", "subcounty", "ward", "facility"),
  organisations = NULL,
  categories = NULL,
  elements = NULL,
  ...
)

Arguments

element_ids

A vector of data element IDs for which to retrieve data. Required.

start_date

The start date to retrieve data. It is required and in the format 'YYYY-MM-dd'.

end_date

The ending date for data retrieval (default is the current date).

level

The desired data granularity: "kenya" (the default), "county", "subcounty", "ward", or "facility".

organisations

A list of organization units in the data. If NULL, downloaded using get_organisation_units_metadata().

categories

A list of categories to include. If NULL, downloaded using get_category_options_metadata().

elements

A list of data elements to include. If NULL, downloaded using get_data_elements_metadata().

...

Other options that can be passed onto .api_get.

Value

A tibble containing breast cancer screening data with the following columns:

kenya - Optional if the level is Kenya.

county - Name of the county. Optional if the level is county, subcounty, ward or facility.

subcounty - Name of the subcounty. Optional if the level is subcounty, ward or facility.

ward - Name of the ward. Optional if the level is ward or facility.

facility - Name of the health facility. Optional if the level facility.

period - The month and year of the data.

fiscal_year- The financial year of the report(July-June Cycle).

year - The calendar year of the report.

month - The month name of the report.

category - The age group category of the report (25-34, 35-39, 40-55, 56-74, or 75+).

category2 - Additional category if available.

element - The data element.

value - The number reported.

See also

Examples

if (FALSE) { # khis_has_cred()
# Clinical Breast Examination data elements
# XEX93uLsAm2 = CBE Abnormal
# cXe64Yk0QMY = CBE Normal
element_id = c('cXe64Yk0QMY', 'XEX93uLsAm2')

# Download data from February 2023 to current date
data <- .get_breast_data(element_ids = element_id, start_date = '2023-02-01')
data
}