Get the Clinical Breast Examination Target Population

get_breast_cbe_target_population() subsets the Kenyan population creates a target population eligible for clinical breast examination: females aged between 25 years and 74 years

Usage

get_breast_cbe_target_population(
  year,
  level = c("kenya", "county", "subcounty")
)

Arguments

year

Year for which to estimate population

level

The desired level of the organization unit hierarchy to retrieve data for: "kenya" (default) , "county" or "subcounty".

Value

A tibble containing the target population for clinical breast examination (CBE)

county - name of the county. Optional if the level is county or subcounty

subcounty - name of the county. Optional if the level if subcounty

target - number to be screened

Details

Implements age group with the min_age set at 25 and max_age at 75 as guided by the Kenya National Cancer Screening Guideline 2018

Examples

# Get the country projection of women to perform CBE for the year 2024
target_population <- get_breast_cbe_target_population(2024)
target_population
#> # A tibble: 1 × 1
#>     target
#>      <dbl>
#> 1 1545577.

# Get the projection for CBE for 2022 by county
target_population <- get_breast_cbe_target_population(2022, level = 'county')
target_population
#> # A tibble: 47 × 2
#>    county          target
#>    <fct>            <dbl>
#>  1 Baringo          9206.
#>  2 Bomet           13021.
#>  3 Bungoma         24429.
#>  4 Busia           13914.
#>  5 Elgeyo Marakwet  6599.
#>  6 Embu            11919.
#>  7 Garissa          9730.
#>  8 Homa Bay        17383.
#>  9 Isiolo           3476.
#> 10 Kajiado         18656.
#> # ℹ 37 more rows