Retrieves cervical cancer screening data with positive results

get_cervical_positive() retrieves cervical cancer screening data with positive results for a specified period from the KHIS API server.

get_cervical_positive(
  start_date,
  end_date = NULL,
  level = c("kenya", "county", "subcounty", "ward", "facility"),
  organisations = NULL,
  categories = NULL,
  elements = NULL,
  khis_session = dynGet("khis_default_session", inherits = TRUE),
  retry = 1,
  verbosity = 0
)

Arguments

start_date

The start date for the data retrieval in the format 'YYYY-MM-dd'.

end_date

The end date for the data retrieval in the format 'YYYY-MM-dd'. Defaults to the current date if not specified.

level

The desired level of the organization unit hierarchy to retrieve data for: "kenya", "county", "subcounty", "ward", or "facility".

organisations

A vector of specific organisation unit IDs to retrieve data for. If NULL, data for all organisation units at the specified level will be retrieved.

categories

A tibble of category options, typically obtained from get_categories(). If NULL, categories will be downloaded automatically.

elements

A tibble of data elements, typically obtained from get_data_elements(). If NULL, elements will be downloaded automatically.

khis_session

The KHIS session object to use (defaults to "khis_default_session"). See ?login_to_khis() for details.

retry

Number of times to retry the API call in case of failure (defaults to 1).

verbosity

Level of information to print during the API call:

0: No output

1: Show headers

2: Show headers and bodies

3: Show headers, bodies, and curl status message

Value

A tibble containing cervical cancer screening data with positive results with the following columns:

See also