Logins into a KHIS system

loginToKHIS provides credentials for making an API call

login_to_khis(
  config_path = NULL,
  config_path_level = "khis",
  username = NULL,
  password = NULL,
  khis_session_name = "khis_default_session",
  khis_session_envir = parent.frame()
)

Arguments

config_path

path to a KHIS config file. If provided, username and password should not be provided.

config_path_level

if there a multiple json entries in the config file, it will default to KHIS

username

KHIS username. If provided config_path must be NULL

password

KHIS password for the username. Optional If not provided you will be prompted to input password

khis_session_name

the variable name for the khisSession object. The default name is d2_default_session and will be used by other cancerscreening functions by default when connecting to KHIS Generally a custom name should only be needed if you need to log into two separate KHIS instances at the same time. If you create a khisSession object with a custom name then this object must be passed to other cancerscreening functions explicitly

khis_session_envir

the environment in which to place the R6 login object, default is the immediate calling environment